+++ src/update-reducer.js
@@ -1,0 +1,12 @@
+const dirty = (state = false, action) => {
+  switch (action.type) {
+    case 'UPDATE':
+      return true
+    case 'DID_UPDATE':
+      return false
+    default:
+      return state
+  }
+}
+
+export default dirty
